Strip CN3SndObjStream
Its entire purpose is to only push chunks of the stream to be played at a time.
While this is efficient, it's not really essential right now, and certainly
does not need to be its own class.

Fundamentally everything behaves essentially the same without, we just use a little
more memory.

With OpenAL we only really need to rework CN3SndObj. Once that's done, we can sort
the threaded frame decoding, so as to avoid pushing the entire buffer at once.